# root/week06/server/src/types.ts
export interface SuccessResponse<T = unknown> {
  isSuccess: true;
  code: string;
  message: string;
  result: T;
}

export interface ErrorResponse {
  isSuccess: false;
  code: string;
  message: string;
}